Add localkube as a bootstrapper
Introduce the --bootstraper or -b flag that toggles the cluster
bootstrapper.  The default bootstrapper (and only bootstrapper at this
time) is localkube.

The bootstrapper interface contains all the functions needed to
bootstrap a kubernetes cluster.  It ingests a KubernetesConfig, which
contains all the configuration for the cluster, as well as a
CommandRunner, which is the interface upon which it will run commands
and transfer files to possibly remotely set up the cluster.

This moves the localkube functions out of cluster.go into their own
package at k8s.io/minikube/pkg/minikube/bootstrapper/localkube.
