+++ pkg/minikube/cluster/cluster_test.go
@@ -18,7 +18,6 @@ package cluster
 
 import (
 	"os"
-	"path/filepath"
 	"strings"
 	"testing"
 
@@ -26,7 +25,6 @@ import (
 	"github.com/docker/machine/libmachine/host"
 	"github.com/docker/machine/libmachine/provision"
 	"github.com/docker/machine/libmachine/state"
-	"k8s.io/minikube/pkg/minikube/bootstrapper"
 	"k8s.io/minikube/pkg/minikube/config"
 	"k8s.io/minikube/pkg/minikube/constants"
 	"k8s.io/minikube/pkg/minikube/tests"
@@ -78,224 +76,6 @@ func TestCreateHost(t *testing.T) {
 
 	if !found {
 		t.Fatalf("Wrong driver name: %v. Should be virtualbox, vmwarefusion, kvm or xhyve.", h.DriverName)
-	}
-}
-
-func TestStartCluster(t *testing.T) {
-	expectedStartCmd, err := GetStartCommand(KubernetesConfig{})
-	if err != nil {
-		t.Fatalf("generating start command: %s", err)
-	}
-
-	cases := []struct {
-		description string
-		startCmd    string
-	}{
-		{
-			description: "start cluster success",
-			startCmd:    expectedStartCmd,
-		},
-		{
-			description: "start cluster failure",
-			startCmd:    "something else",
-		},
-	}
-
-	for _, test := range cases {
-		t.Run(test.description, func(t *testing.T) {
-			t.Parallel()
-			f := bootstrapper.NewFakeCommandRunner()
-			f.SetCommandToOutput(map[string]string{test.startCmd: "ok"})
-			err := StartCluster(f, KubernetesConfig{})
-			if err != nil && test.startCmd == expectedStartCmd {
-				t.Errorf("Error starting cluster: %s", err)
-			}
-		})
-	}
-}
-
-func TestUpdateCluster(t *testing.T) {
-	defaultCfg := KubernetesConfig{
-		KubernetesVersion: constants.DefaultKubernetesVersion,
-	}
-	defaultAddons := []string{
-		"deploy/addons/kube-dns/kube-dns-cm.yaml",
-		"deploy/addons/kube-dns/kube-dns-svc.yaml",
-		"deploy/addons/addon-manager.yaml",
-		"deploy/addons/dashboard/dashboard-rc.yaml",
-		"deploy/addons/dashboard/dashboard-svc.yaml",
-		"deploy/addons/storageclass/storageclass.yaml",
-		"deploy/addons/kube-dns/kube-dns-controller.yaml",
-	}
-	cases := []struct {
-		description   string
-		k8s           KubernetesConfig
-		expectedFiles []string
-		shouldErr     bool
-	}{
-		{
-			description:   "transfer localkube correct",
-			k8s:           defaultCfg,
-			expectedFiles: []string{"out/localkube"},
-		},
-		{
-			description:   "addons are transferred",
-			k8s:           defaultCfg,
-			expectedFiles: defaultAddons,
-		},
-		{
-			description: "no localkube version",
-			k8s:         KubernetesConfig{},
-			shouldErr:   true,
-		},
-	}
-
-	for _, test := range cases {
-		t.Run(test.description, func(t *testing.T) {
-			t.Parallel()
-			f := bootstrapper.NewFakeCommandRunner()
-			err := UpdateCluster(f, test.k8s)
-			if err != nil && !test.shouldErr {
-				t.Errorf("Error updating cluster: %s", err)
-				return
-			}
-			if err == nil && test.shouldErr {
-				t.Error("Didn't get error, but expected to")
-				return
-			}
-			for _, expectedFile := range test.expectedFiles {
-				_, err := f.GetFileToContents(expectedFile)
-				if err != nil {
-					t.Errorf("Expected file %s, but was not present", expectedFile)
-				}
-			}
-		})
-	}
-}
-
-func TestGetLocalkubeStatus(t *testing.T) {
-	cases := []struct {
-		description    string
-		statusCmdMap   map[string]string
-		expectedStatus string
-		shouldErr      bool
-	}{
-		{
-			description:    "get status running",
-			statusCmdMap:   map[string]string{localkubeStatusCommand: "Running"},
-			expectedStatus: "Running",
-		},
-		{
-			description:    "get status stopped",
-			statusCmdMap:   map[string]string{localkubeStatusCommand: "Stopped"},
-			expectedStatus: "Stopped",
-		},
-		{
-			description:  "get status unknown status",
-			statusCmdMap: map[string]string{localkubeStatusCommand: "Recalculating..."},
-			shouldErr:    true,
-		},
-		{
-			description:  "get status error",
-			statusCmdMap: map[string]string{"a": "b"},
-			shouldErr:    true,
-		},
-	}
-
-	for _, test := range cases {
-		t.Run(test.description, func(t *testing.T) {
-			t.Parallel()
-			f := bootstrapper.NewFakeCommandRunner()
-			f.SetCommandToOutput(test.statusCmdMap)
-			actualStatus, err := GetLocalkubeStatus(f)
-			if err != nil && !test.shouldErr {
-				t.Errorf("Error getting localkube status: %s", err)
-				return
-			}
-			if err == nil && test.shouldErr {
-				t.Error("Didn't get error, but expected to")
-				return
-			}
-			if test.expectedStatus != actualStatus {
-				t.Errorf("Expected status: %s, Actual status: %s", test.expectedStatus, actualStatus)
-			}
-		})
-	}
-}
-
-func TestGetHostLogs(t *testing.T) {
-	logs, err := GetLogsCommand(false)
-	if err != nil {
-		t.Fatalf("Error getting logs command: %s", err)
-	}
-	logsf, err := GetLogsCommand(true)
-	if err != nil {
-		t.Fatalf("Error gettings logs -f command: %s", err)
-	}
-
-	cases := []struct {
-		description string
-		logsCmdMap  map[string]string
-		follow      bool
-		shouldErr   bool
-	}{
-		{
-			description: "get logs correct",
-			logsCmdMap:  map[string]string{logs: "fee"},
-		},
-		{
-			description: "follow logs correct",
-			logsCmdMap:  map[string]string{logsf: "fi"},
-			follow:      true,
-		},
-		{
-			description: "get logs incorrect",
-			logsCmdMap:  map[string]string{"fo": "fum"},
-			shouldErr:   true,
-		},
-	}
-
-	for _, test := range cases {
-		t.Run(test.description, func(t *testing.T) {
-			t.Parallel()
-			f := bootstrapper.NewFakeCommandRunner()
-			f.SetCommandToOutput(test.logsCmdMap)
-			_, err := GetHostLogs(f, test.follow)
-			if err != nil && !test.shouldErr {
-				t.Errorf("Error getting localkube logs: %s", err)
-				return
-			}
-			if err == nil && test.shouldErr {
-				t.Error("Didn't get error, but expected to")
-				return
-			}
-		})
-	}
-}
-
-func TestSetupCerts(t *testing.T) {
-	tempDir := tests.MakeTempDir()
-	defer os.RemoveAll(tempDir)
-
-	f := bootstrapper.NewFakeCommandRunner()
-	k8s := KubernetesConfig{
-		APIServerName: constants.APIServerName,
-		DNSDomain:     constants.ClusterDNSDomain,
-	}
-
-	var filesToBeTransferred []string
-	for _, cert := range certs {
-		filesToBeTransferred = append(filesToBeTransferred, filepath.Join(constants.GetMinipath(), cert))
-	}
-
-	if err := SetupCerts(f, k8s); err != nil {
-		t.Fatalf("Error starting cluster: %s", err)
-	}
-	for _, cert := range filesToBeTransferred {
-		_, err := f.GetFileToContents(cert)
-		if err != nil {
-			t.Errorf("Cert not generated: %s", cert)
-		}
 	}
 }
 
@@ -607,48 +387,5 @@ func TestCreateSSHShell(t *testing.T) {
 
 	if !s.IsSessionRequested() {
 		t.Fatalf("Expected ssh session to be run")
-	}
-}
-
-func TestIsLocalkubeCached(t *testing.T) {
-	tempDir := tests.MakeTempDir()
-	defer os.RemoveAll(tempDir)
-
-	inputArr := [...]string{
-		"v1.3.3",
-		"1.3.0",
-		"http://test-url.localkube.com/localkube-binary",
-		"file:///test/dir/to/localkube-binary",
-	}
-
-	localkubeCacher := localkubeCacher{
-		k8sConf: KubernetesConfig{},
-	}
-
-	inner := func(input string) {
-		localkubeCacher.k8sConf = KubernetesConfig{
-			KubernetesVersion: input,
-		}
-		if localkubeCacher.isLocalkubeCached() {
-			t.Errorf("IsLocalKubeCached returned true even though %s was not cached",
-				localkubeCacher.getLocalkubeCacheFilepath())
-			return
-		}
-
-		f, err := os.Create(localkubeCacher.getLocalkubeCacheFilepath())
-		if err != nil {
-			t.Errorf("failed to create dummy cache file: %v", err)
-			return
-		}
-		f.Close()
-		defer os.Remove(f.Name())
-		if !localkubeCacher.isLocalkubeCached() {
-			t.Errorf("IsLocalKubeCached returned false even though %s was cached",
-				localkubeCacher.getLocalkubeCacheFilepath())
-		}
-
-	}
-	for _, input := range inputArr {
-		inner(input)
 	}
 }
